+++ infra/bots/task_drivers/perf_puppeteer_render_skps/perf_puppeteer_render_skps.go
@@ -26,6 +26,8 @@ import (
 	"go.skia.org/infra/task_driver/go/td"
 )
 
+const perfKeyWebGLVersion = "webgl_version"
+
 func main() {
 	var (
 		// Required properties for this task.
@@ -42,6 +44,7 @@ func main() {
 		modelTrace         = flag.String("model_trace", "", "Description of host machine.")
 		cpuOrGPUTrace      = flag.String("cpu_or_gpu_trace", "", "If this is a CPU or GPU configuration.")
 		cpuOrGPUValueTrace = flag.String("cpu_or_gpu_value_trace", "", "The hardware of this CPU/GPU")
+		webGLVersion       = flag.String("webgl_version", "", "Major WebGL version to use when creating gl drawing context. 1 or 2")
 
 		// Flags that may be required for certain configs
 		canvaskitBinPath = flag.String("canvaskit_bin_path", "", "The location of a canvaskit.js and canvaskit.wasm")
@@ -57,10 +60,11 @@ func main() {
 	defer td.EndRun(ctx)
 
 	keys := map[string]string{
-		"os":               *osTrace,
-		"model":            *modelTrace,
-		perfKeyCpuOrGPU:    *cpuOrGPUTrace,
-		"cpu_or_gpu_value": *cpuOrGPUValueTrace,
+		"os":                *osTrace,
+		"model":             *modelTrace,
+		perfKeyCpuOrGPU:     *cpuOrGPUTrace,
+		"cpu_or_gpu_value":  *cpuOrGPUValueTrace,
+		perfKeyWebGLVersion: *webGLVersion,
 	}
 
 	outputWithoutResults, err := makePerfObj(*gitHash, *taskID, os.Getenv("SWARMING_BOT_ID"), keys)
@@ -188,6 +192,9 @@ func benchSKPs(ctx context.Context, perf perfJSONFormat, benchmarkPath, canvaski
 			}
 			if perf.Key[perfKeyCpuOrGPU] != "CPU" {
 				args = append(args, "--use_gpu")
+				if perf.Key[perfKeyWebGLVersion] == "1" {
+					args = append(args, "--query_params webgl1")
+				}
 			}
 			_, err := exec.RunCwd(ctx, benchmarkPath, args...)
 			if err != nil {
@@ -244,6 +251,9 @@ func processSKPData(ctx context.Context, perf perfJSONFormat, benchmarkPath, out
 			config := "software"
 			if perf.Key[perfKeyCpuOrGPU] != "CPU" {
 				config = "webgl2"
+				if perf.Key[perfKeyWebGLVersion] == "1" {
+					config = "webgl1"
+				}
 			}
 
 			b, err := os_steps.ReadFile(ctx, skp)