+++ modules/canvaskit/externs.js
@@ -556,6 +556,7 @@ var CanvasKit = {
     sampleCnt: function() {},
     reportBackendTypeIsGPU: function() {},
     grContext: {},
+    openGLversion: {},
 
     // private API
     _flush: function() {},
@@ -1055,3 +1056,11 @@ var DOMMatrix = {
 
 // Not sure why this is needed - might be a bug in emsdk that this isn't properly declared.
 function loadWebAssemblyModule() {};
+
+// This is a part of emscripten's webgl glue code. Preserving this attribute is necessary
+// to override it in the puppeteer tests
+var LibraryEGL = {
+  contextAttributes: {
+    majorVersion: {}
+  }
+}