+++ tools/perf-canvaskit-puppeteer/benchmark.js
@@ -1,10 +1,9 @@
 // Shared benchmarking functions such as surface creation, measurement, publishing to perf.skia.org
 
-// TODO(kjlubick) make this configurable to return a WEBGL 1 or WEBGL 2 surface.
-function getSurface(CanvasKit) {
+function getSurface(CanvasKit, webglversion) {
   let surface;
   if (window.location.hash.indexOf('gpu') !== -1) {
-    surface = CanvasKit.MakeWebGLCanvasSurface('anim');
+    surface = CanvasKit.MakeWebGLCanvasSurface('anim', null /* colorspace */, {'majorVersion': webglversion});
     if (!surface) {
       window._error = 'Could not make GPU surface';
       return null;
@@ -14,6 +13,10 @@ function getSurface(CanvasKit) {
     // to CPU and add a ck-replaced class to the canvas element.
     if (c.classList.contains('ck-replaced')) {
       window._error = 'fell back to CPU';
+      return null;
+    }
+    if (webglversion != surface.openGLversion) {
+      window._error = 'Want WebGL version '+webglversion+' but got '+surface.openGLversion;
       return null;
     }
   } else {