+++ modules/canvaskit/CHANGELOG.md
@@ -34,7 +34,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    MakeSkVertices is the flat Uint32Array.
  - Color arrays may also be objects created with CanvasKit.Malloc
  - renamed `reportBackendType` to `reportBackendTypeIsGPU` and made it return a boolean
-
+ - `MakeWebGLCanvasSurface` can now accept an optional dictionary of WebGL context attributes that
+   can be used to override default attributes.
 ### Fixed
  - `TextStyle.color` can correctly be a Malloc'd Float32Array.